crate an outer TransportWriter type that wraps real/mock and provides decoding to both
